final refactors and docs
